Refactoring: - moved HdEngine to engine/engine.py - moved handler functions to engine/handlers.py
